% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scal_g1m.R
\name{scal_g1m}
\alias{scal_g1m}
\title{Calculates sample size for estimating a mean}
\usage{
scal_g1m(sigma, d, alpha)
}
\arguments{
\item{d}{precision or margin of the error for the outcome}

\item{alpha}{level for the confidence interval}

\item{\sigma}{expected variability of the outcome}
}
\value{
output Sample size
}
\description{
The function requires 1) desired precision 2) expected variability in outcome and 3)
level of confidence as required input.
}
\examples{
scal_g1m(10, 3, 0.05)


}
\keyword{,}
\keyword{means,}
\keyword{precision}

