% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scal_g2m.R
\name{scal_g2m}
\alias{scal_g2m}
\title{Calculates sample size for comparing two means}
\usage{
scal_g2m(mu1, mu2, sd1, sd2, alpha, power)
}
\arguments{
\item{alpha}{desired level of significance}

\item{power}{desired level of power}

\item{\mu_{1}}{expected mean in group1}

\item{\mu_{2}}{expected mean in group2}

\item{\sd_{1}}{expected sd in group1}

\item{\sd_{2}}{expected sd in group2}
}
\value{
output Sample size
}
\description{
The function requires 1) expected means & sds in two groups 2) desired significance level 3)
level of power as required input.
}
\examples{
scal_g2m(50, 70, 20, 20, 0.05, 0.80)


}
\keyword{groups,}
\keyword{means,two}
\keyword{power}

