% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scal_g1p.R
\name{scal_g1p}
\alias{scal_g1p}
\title{Calculates sample size for estimating a proportion}
\usage{
scal_g1p(p, d, alpha)
}
\arguments{
\item{p}{expected or assummed proportion of the outcome}

\item{d}{precision or margin of the error for the outcome}

\item{alpha}{level for the confidence interval}
}
\value{
Sample size
}
\description{
The function requires 1) assummed proportion 2) precision and 3)
level of confidence as required input.
}
\examples{
scal_g1p(p=0.5, d=0.05, alpha=0.05)
}
\keyword{precision}
\keyword{prevalence,}
\keyword{proportions,}

